Improved portability
Ignore-this: c41b037b3b89745a6728afd5ee00b0b3

For fussy compilers such as xlc.

darcs-hash:153276f04d49aaae9a4d8cd9705e127210e7e959

+++ src/utils.h
@@ -103,7 +103,7 @@ void timer_print (timer t, int i, long tnc)
   }
   printf ("# " GRIDNAME 
 	  ", %d steps, %g CPU, %.4g real, %.3g points.step/s, %d var\n",
-	  i, cpu, real, tnc/real, datasize/sizeof(double));
+	  i, cpu, real, tnc/real, (int) (datasize/sizeof(double)));
 }
 
 typedef struct {
